report and metric linking

+++ src/main/java/edu/university/ecs/lab/impact/metrics/services/SystemMetricsService.java
@@ -77,7 +77,7 @@ public class SystemMetricsService {
             }
         }
 
-        return links.size() / (microserviceMap.values().size() * microserviceMap.values().size()) - microserviceMap.values().size();
+        return (double) links.size() / (microserviceMap.values().size() * microserviceMap.values().size()) - microserviceMap.values().size();
     }
 
     private RestCall updateRestCallDest(RestCall restCall) {
